Improve code portability
Minor additions/modifications required to existing code-base for compatibility with espressif esp32 compiler.

+++ src/usb-cdc.c
@@ -1,6 +1,7 @@
 #include <hardware/watchdog.h>
 #include <pico/bootrom.h>
 #include <stdlib.h>
+#include <string.h>
 #include <tusb.h>
 
 #include "cec-log.h"
@@ -200,8 +201,8 @@ static int show_stats_tasks(void) {
   cdc_printfln("%-13s | %-10s | %-10s", "task", "priority", "stack (min)");
 
   for (UBaseType_t i = 0; i < n; i++) {
-    cdc_printfln("%-13s | %-10lu | %-10lu", status[i].pcTaskName, status[i].uxCurrentPriority,
-                 status[i].usStackHighWaterMark);
+    cdc_printfln("%-13s | %-10lu | %-10lu", status[i].pcTaskName, (uint32_t)status[i].uxCurrentPriority,
+                 (uint32_t)status[i].usStackHighWaterMark);
   }
 
   return 0;
